requirements.txt
# CipherBuster v2.0 - Dépendances

# Crypto
pycryptodome>=3.18.0
gmpy2>=2.1.5

# Interface CLI
rich>=13.0.0
termcolor>=2.3.0

# Utilitaires
sympy>=1.12

# Base de données
factordb-pycli>=1.0.0

# Cryptographie avancée
cryptography>=41.0.0

# Optionnel mais recommandé
# pytest>=7.4.0  # Pour tests unitaires
# pytest-cov>=4.1.0  # Coverage